vapi_agent_forge/backend/requirements.txt
# Vapi Agent Forge Requirements  
# Python 3.11+ Compatible

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
PyYAML==6.0.1
python-multipart==0.0.6

# Additional utilities
requests==2.31.0 